feat(auth): correct commonjs exports for latest authjs
